Moved Providers implementation files to the Providers folder of the Infra assembly. While this introduces a hard dependency between client side UI projects and Infra, inc un-needed anciliairy nuget dependencies, it can be split into multiple infra assemblies in future, mitigating this impact.
