+++ src/AdminAssistant.Blazor/Client/AdminAssistant.Blazor.Client.csproj
@@ -42,6 +42,7 @@
   </ItemGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\..\AdminAssistant.Infra\AdminAssistant.Infra.csproj" />
     <ProjectReference Include="..\..\AdminAssistant.UI\AdminAssistant.UI.csproj" />
   </ItemGroup>
