+++ src/AdminAssistant.Desktop/WPF/AdminAssistant.WPF/AdminAssistant.WPF.csproj
@@ -41,6 +41,7 @@
   </ItemGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\..\..\AdminAssistant.Infra\AdminAssistant.Infra.csproj" />
     <ProjectReference Include="..\..\..\AdminAssistant.UI\AdminAssistant.UI.csproj" />
   </ItemGroup>
